rename: view 파일들 view 패키지로 이동
